GNG-484: OnThisPage concerns are broken up into Editing, UI, and Command files
